Refactor long command to node script
