+++ package.json
@@ -8,7 +8,7 @@
   },
   "scripts": {
     "autocrop-images": "babel-node tools/autocropImages",
-    "dev": "yarn yaml2json && yarn copy-landscape && next dev",
+    "dev": "yarn yaml2json && yarn prepare-landscape && next dev",
     "landscapes": "babel-node tools/landscapes.js",
     "lint": "esw webpack.config.* src tools --color",
     "lint:watch": "yarn lint -- --watch",
@@ -32,10 +32,10 @@
     "render-landscape": "babel-node tools/renderLandscape",
     "funding": "babel-node tools/fundingForMasterBranch",
     "copy-dist": "cp -r _headers out && ([ -z \"${PROJECT_NAME:-}\" ] || mkdir \"$PROJECT_PATH\"/dist ) && mv out \"$PROJECT_PATH\"/dist/${PROJECT_NAME:-}",
-    "copy-landscape": "rm -rf public && mkdir -p public && cp -r \"$PROJECT_PATH\"/images \"$PROJECT_PATH\"/cached_logos public && mv public/cached_logos public/logos && babel-node tools/parseSettings.js",
+    "prepare-landscape": "babel-node tools/prepareLandscape.js",
     "setup-robots": "babel-node tools/sitemap && babel-node tools/addRobots",
     "quick-build": "babel-node tools/build.js && yarn setup-robots && yarn copy-dist ",
-    "build": "((yarn fetch && yarn copy-landscape && yarn build-next && yarn setup-robots && yarn remove-dist && yarn copy-dist && yarn stop-old-ci && yarn start-ci && babel-node tools/parallelWithRetry integration-test check-landscape render-landscape funding) || (yarn stop-old-ci && false)) && babel-node tools/parallelWithRetry integration-test && yarn stop-ci",
+    "build": "((yarn fetch && yarn prepare-landscape && yarn build-next && yarn setup-robots && yarn remove-dist && yarn copy-dist && yarn stop-old-ci && yarn start-ci && babel-node tools/parallelWithRetry integration-test check-landscape render-landscape funding) || (yarn stop-old-ci && false)) && babel-node tools/parallelWithRetry integration-test && yarn stop-ci",
     "build-next": "next build && next export",
     "show-report": "open dist/report.html",
     "analyze-bundle": "babel-node ./tools/analyzeBundle.js",
